+++ AtelierApp/UI/PageListTextile.xaml.cs
@@ -36,20 +36,33 @@ namespace AtelierApp.UI
         private void btnDelete_Click(object sender, RoutedEventArgs e)
         {
             var textileForRemoving = DGridTextile.SelectedItems.Cast<Textile>().ToList();
+            int i = 0;
             if (MessageBox.Show($"Вы точно хотите удалить следующие {textileForRemoving.Count()} элементов?", "Внимание!",
                 MessageBoxButton.YesNo, MessageBoxImage.Question) == MessageBoxResult.Yes)
             {
-                try
+                foreach (Textile item in textileForRemoving)
                 {
-                    AtelierBaseEntities.GetContext().Textile.RemoveRange(textileForRemoving);
-                    AtelierBaseEntities.GetContext().SaveChanges();
-                    MessageBox.Show("Данные удалены!", "Успешно!", MessageBoxButton.OK, MessageBoxImage.Information);
-                    textile = AtelierBaseEntities.GetContext().Textile.ToList();
-                    DGridTextile.ItemsSource = textile;
+                    if (AtelierBaseEntities.GetContext().Order.Where(o => o.IdTextile == item.Id).ToList().FirstOrDefault() != null)
+                    {
+                        MessageBox.Show("С удаляемыми данными имеются связанные записи в других таблицах.", "Удаление отменено!", MessageBoxButton.OK, MessageBoxImage.Error);
+                        i = 1;
+                        break;
+                    }
                 }
-                catch (Exception ex)
+                if (i == 0)
                 {
-                    MessageBox.Show("С удаляемыми данными имеются связанные записи в других таблицах.", "Удаление отменено!", MessageBoxButton.OK, MessageBoxImage.Error);
+                    try
+                    {
+                        AtelierBaseEntities.GetContext().Textile.RemoveRange(textileForRemoving);
+                        AtelierBaseEntities.GetContext().SaveChanges();
+                        MessageBox.Show("Данные удалены!", "Успешно!", MessageBoxButton.OK, MessageBoxImage.Information);
+                        textile = AtelierBaseEntities.GetContext().Textile.ToList();
+                        DGridTextile.ItemsSource = textile;
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show("При удалении данных возникли неполадки!", "Удаление отменено!", MessageBoxButton.OK, MessageBoxImage.Error);
+                    }
                 }
             }
         }
